Extract widgets from settings feature
- Extract AgentCard, TemplateCard, AgentList, TemplateList widgets
- Extract VoiceSettingsSection and CreateAgentDialog widgets
- Create widgets directory structure for settings feature
- Update agent_management_screen.dart to use extracted widgets
- Clean up unused imports and remove old widget classes
- Maintain widget extraction pattern consistency across features
