+++ lib/features/settings/presentation/agent_management_screen.dart
@@ -1,13 +1,15 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import '../bloc/agent_bloc.dart';
-import '../bloc/agent_events.dart';
 import '../bloc/agent_states.dart';
 import '../models/agent_model.dart';
 import '../../../shared/widgets/loading_indicator.dart';
-import '../../../shared/widgets/voice_components.dart';
-import '../../voice/bloc/voice_bloc.dart';
-import '../../voice/bloc/voice_state.dart';
+import '../widgets/agent_empty_state.dart';
+import '../widgets/agent_error_message.dart';
+import '../widgets/agent_list.dart';
+import '../widgets/template_list.dart';
+import '../widgets/voice_settings_section.dart';
+import '../widgets/create_agent_dialog.dart';
 
 class AgentManagementScreen extends StatelessWidget {
   const AgentManagementScreen({super.key});
@@ -31,73 +33,20 @@ class AgentManagementScreen extends StatelessWidget {
           }
 
           if (state is AgentError) {
-            return Center(
-              child: Column(
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: [
-                  Icon(
-                    Icons.error_outline,
-                    size: 64,
-                    color: Theme.of(context).colorScheme.error,
-                  ),
-                  const SizedBox(height: 16),
-                  Text(
-                    'Error',
-                    style: Theme.of(context).textTheme.headlineSmall,
-                  ),
-                  const SizedBox(height: 8),
-                  Text(
-                    state.message,
-                    textAlign: TextAlign.center,
-                    style: Theme.of(context).textTheme.bodyMedium,
-                  ),
-                ],
-              ),
-            );
+            return AgentErrorMessage(message: state.message);
           }
 
           if (state is AgentLoaded) {
             return _buildAgentList(context, state);
           }
 
-          return _buildEmptyState(context);
+          return AgentEmptyState(
+              onCreateAgent: () => _showCreateAgentDialog(context));
         },
       ),
       floatingActionButton: FloatingActionButton(
         onPressed: () => _showCreateAgentDialog(context),
         child: const Icon(Icons.add),
-      ),
-    );
-  }
-
-  Widget _buildEmptyState(BuildContext context) {
-    return Center(
-      child: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: [
-          Icon(
-            Icons.smart_toy,
-            size: 64,
-            color: Theme.of(context).colorScheme.primary,
-          ),
-          const SizedBox(height: 16),
-          Text(
-            'No Custom Agents',
-            style: Theme.of(context).textTheme.headlineSmall,
-          ),
-          const SizedBox(height: 8),
-          Text(
-            'Create custom AI agents for different tasks',
-            textAlign: TextAlign.center,
-            style: Theme.of(context).textTheme.bodyMedium,
-          ),
-          const SizedBox(height: 24),
-          ElevatedButton.icon(
-            onPressed: () => _showCreateAgentDialog(context),
-            icon: const Icon(Icons.add),
-            label: const Text('Create Agent'),
-          ),
-        ],
       ),
     );
   }
@@ -131,556 +80,25 @@ class AgentManagementScreen extends StatelessWidget {
 
   Widget _buildMyAgents(
       BuildContext context, List<AgentModel> agents, AgentModel? activeAgent) {
-    if (agents.isEmpty) {
-      return _buildEmptyState(context);
-    }
-
-    return ListView.builder(
-      padding: const EdgeInsets.all(16),
-      itemCount: agents.length,
-      itemBuilder: (context, index) {
-        final agent = agents[index];
-        return _AgentCard(
-          agent: agent,
-          isActive: activeAgent?.id == agent.id,
-        );
-      },
+    return AgentList(
+      agents: agents,
+      activeAgent: activeAgent,
+      onCreateAgent: () => _showCreateAgentDialog(context),
     );
   }
 
   Widget _buildTemplates(BuildContext context, List<AgentModel> templates) {
-    return ListView.builder(
-      padding: const EdgeInsets.all(16),
-      itemCount: templates.length,
-      itemBuilder: (context, index) {
-        final template = templates[index];
-        return _TemplateCard(template: template);
-      },
-    );
+    return TemplateList(templates: templates);
   }
 
   Widget _buildVoiceSettings(BuildContext context) {
-    return SingleChildScrollView(
-      padding: const EdgeInsets.all(16),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Text(
-            'Voice Settings',
-            style: Theme.of(context).textTheme.headlineSmall?.copyWith(
-                  fontWeight: FontWeight.bold,
-                ),
-          ),
-          const SizedBox(height: 16),
-
-          // Voice output toggle
-          VoiceOutputToggle(),
-
-          const SizedBox(height: 16),
-
-          // Voice settings panel
-          VoiceSettingsPanel(),
-
-          const SizedBox(height: 16),
-
-          // Voice status indicator
-          BlocBuilder<VoiceBloc, VoiceState>(
-            builder: (context, state) {
-              if (state is VoiceError) {
-                return Card(
-                  color: Theme.of(context).colorScheme.errorContainer,
-                  child: Padding(
-                    padding: const EdgeInsets.all(16),
-                    child: Row(
-                      children: [
-                        Icon(
-                          Icons.error_outline,
-                          color: Theme.of(context).colorScheme.onErrorContainer,
-                        ),
-                        const SizedBox(width: 8),
-                        Expanded(
-                          child: Text(
-                            state.message,
-                            style: TextStyle(
-                              color: Theme.of(context)
-                                  .colorScheme
-                                  .onErrorContainer,
-                            ),
-                          ),
-                        ),
-                      ],
-                    ),
-                  ),
-                );
-              }
-
-              if (state is VoiceNotAvailable) {
-                return Card(
-                  color: Theme.of(context).colorScheme.errorContainer,
-                  child: Padding(
-                    padding: const EdgeInsets.all(16),
-                    child: Row(
-                      children: [
-                        Icon(
-                          Icons.mic_off,
-                          color: Theme.of(context).colorScheme.onErrorContainer,
-                        ),
-                        const SizedBox(width: 8),
-                        Expanded(
-                          child: Text(
-                            'Voice features not available: ${state.reason}',
-                            style: TextStyle(
-                              color: Theme.of(context)
-                                  .colorScheme
-                                  .onErrorContainer,
-                            ),
-                          ),
-                        ),
-                      ],
-                    ),
-                  ),
-                );
-              }
-
-              return const SizedBox.shrink();
-            },
-          ),
-        ],
-      ),
-    );
+    return const VoiceSettingsSection();
   }
 
   void _showCreateAgentDialog(BuildContext context) {
-    final nameController = TextEditingController();
-    final descriptionController = TextEditingController();
-    final promptController = TextEditingController();
-
     showDialog(
       context: context,
-      builder: (context) => AlertDialog(
-        title: const Text('Create Custom Agent'),
-        content: SingleChildScrollView(
-          child: Column(
-            mainAxisSize: MainAxisSize.min,
-            children: [
-              TextField(
-                controller: nameController,
-                decoration: const InputDecoration(
-                  labelText: 'Agent Name',
-                  border: OutlineInputBorder(),
-                ),
-                autofocus: true,
-              ),
-              const SizedBox(height: 16),
-              TextField(
-                controller: descriptionController,
-                decoration: const InputDecoration(
-                  labelText: 'Description',
-                  border: OutlineInputBorder(),
-                ),
-                maxLines: 2,
-              ),
-              const SizedBox(height: 16),
-              TextField(
-                controller: promptController,
-                decoration: const InputDecoration(
-                  labelText: 'System Prompt',
-                  border: OutlineInputBorder(),
-                  hintText: 'Define the agent\'s behavior and capabilities...',
-                ),
-                maxLines: 4,
-              ),
-            ],
-          ),
-        ),
-        actions: [
-          TextButton(
-            onPressed: () => Navigator.pop(context),
-            child: const Text('Cancel'),
-          ),
-          ElevatedButton(
-            onPressed: () {
-              if (nameController.text.isNotEmpty &&
-                  descriptionController.text.isNotEmpty &&
-                  promptController.text.isNotEmpty) {
-                context.read<AgentBloc>().add(
-                      CreateAgent(
-                        name: nameController.text,
-                        description: descriptionController.text,
-                        systemPrompt: promptController.text,
-                      ),
-                    );
-                Navigator.pop(context);
-              }
-            },
-            child: const Text('Create'),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class _AgentCard extends StatelessWidget {
-  final AgentModel agent;
-  final bool isActive;
-
-  const _AgentCard({
-    required this.agent,
-    required this.isActive,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Card(
-      margin: const EdgeInsets.only(bottom: 12),
-      child: ListTile(
-        leading: Container(
-          padding: const EdgeInsets.all(8),
-          decoration: BoxDecoration(
-            color: Theme.of(context).colorScheme.primaryContainer,
-            borderRadius: BorderRadius.circular(8),
-          ),
-          child: Icon(
-            Icons.smart_toy,
-            color: Theme.of(context).colorScheme.onPrimaryContainer,
-          ),
-        ),
-        title: Row(
-          children: [
-            Expanded(
-              child: Text(
-                agent.name,
-                style: Theme.of(context).textTheme.titleMedium,
-              ),
-            ),
-            if (isActive)
-              Container(
-                padding: const EdgeInsets.symmetric(horizontal: 8, vertical: 2),
-                decoration: BoxDecoration(
-                  color: Theme.of(context).colorScheme.primary,
-                  borderRadius: BorderRadius.circular(12),
-                ),
-                child: Text(
-                  'Active',
-                  style: Theme.of(context).textTheme.bodySmall?.copyWith(
-                        color: Theme.of(context).colorScheme.onPrimary,
-                      ),
-                ),
-              ),
-          ],
-        ),
-        subtitle: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Text(
-              agent.description,
-              style: Theme.of(context).textTheme.bodyMedium,
-            ),
-            const SizedBox(height: 4),
-            Text(
-              'Created: ${_formatDate(agent.createdAt)}',
-              style: Theme.of(context).textTheme.bodySmall,
-            ),
-          ],
-        ),
-        trailing: PopupMenuButton<String>(
-          onSelected: (value) => _handleMenuAction(context, value),
-          itemBuilder: (context) => [
-            if (!isActive)
-              const PopupMenuItem(
-                value: 'activate',
-                child: Row(
-                  children: [
-                    Icon(Icons.play_arrow),
-                    SizedBox(width: 8),
-                    Text('Activate'),
-                  ],
-                ),
-              ),
-            const PopupMenuItem(
-              value: 'edit',
-              child: Row(
-                children: [
-                  Icon(Icons.edit),
-                  SizedBox(width: 8),
-                  Text('Edit'),
-                ],
-              ),
-            ),
-            const PopupMenuItem(
-              value: 'export',
-              child: Row(
-                children: [
-                  Icon(Icons.download),
-                  SizedBox(width: 8),
-                  Text('Export'),
-                ],
-              ),
-            ),
-            const PopupMenuItem(
-              value: 'delete',
-              child: Row(
-                children: [
-                  Icon(Icons.delete, color: Colors.red),
-                  SizedBox(width: 8),
-                  Text('Delete', style: TextStyle(color: Colors.red)),
-                ],
-              ),
-            ),
-          ],
-        ),
-        onTap: () => _showAgentDetails(context),
-      ),
-    );
-  }
-
-  String _formatDate(DateTime date) {
-    return '${date.day}/${date.month}/${date.year}';
-  }
-
-  void _handleMenuAction(BuildContext context, String action) {
-    switch (action) {
-      case 'activate':
-        context.read<AgentBloc>().add(SetActiveAgent(agent.id));
-        break;
-      case 'edit':
-        _showEditDialog(context);
-        break;
-      case 'export':
-        context.read<AgentBloc>().add(ExportAgent(agent.id));
-        break;
-      case 'delete':
-        _showDeleteDialog(context);
-        break;
-    }
-  }
-
-  void _showAgentDetails(BuildContext context) {
-    showDialog(
-      context: context,
-      builder: (context) => AlertDialog(
-        title: Text(agent.name),
-        content: SingleChildScrollView(
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            mainAxisSize: MainAxisSize.min,
-            children: [
-              Text(
-                'Description',
-                style: Theme.of(context).textTheme.titleSmall,
-              ),
-              const SizedBox(height: 4),
-              Text(agent.description),
-              const SizedBox(height: 16),
-              Text(
-                'System Prompt',
-                style: Theme.of(context).textTheme.titleSmall,
-              ),
-              const SizedBox(height: 4),
-              Text(agent.systemPrompt),
-            ],
-          ),
-        ),
-        actions: [
-          TextButton(
-            onPressed: () => Navigator.pop(context),
-            child: const Text('Close'),
-          ),
-        ],
-      ),
-    );
-  }
-
-  void _showEditDialog(BuildContext context) {
-    final nameController = TextEditingController(text: agent.name);
-    final descriptionController =
-        TextEditingController(text: agent.description);
-    final promptController = TextEditingController(text: agent.systemPrompt);
-
-    showDialog(
-      context: context,
-      builder: (context) => AlertDialog(
-        title: const Text('Edit Agent'),
-        content: SingleChildScrollView(
-          child: Column(
-            mainAxisSize: MainAxisSize.min,
-            children: [
-              TextField(
-                controller: nameController,
-                decoration: const InputDecoration(
-                  labelText: 'Agent Name',
-                  border: OutlineInputBorder(),
-                ),
-              ),
-              const SizedBox(height: 16),
-              TextField(
-                controller: descriptionController,
-                decoration: const InputDecoration(
-                  labelText: 'Description',
-                  border: OutlineInputBorder(),
-                ),
-                maxLines: 2,
-              ),
-              const SizedBox(height: 16),
-              TextField(
-                controller: promptController,
-                decoration: const InputDecoration(
-                  labelText: 'System Prompt',
-                  border: OutlineInputBorder(),
-                ),
-                maxLines: 4,
-              ),
-            ],
-          ),
-        ),
-        actions: [
-          TextButton(
-            onPressed: () => Navigator.pop(context),
-            child: const Text('Cancel'),
-          ),
-          ElevatedButton(
-            onPressed: () {
-              if (nameController.text.isNotEmpty &&
-                  descriptionController.text.isNotEmpty &&
-                  promptController.text.isNotEmpty) {
-                context.read<AgentBloc>().add(
-                      UpdateAgent(
-                        id: agent.id,
-                        name: nameController.text,
-                        description: descriptionController.text,
-                        systemPrompt: promptController.text,
-                      ),
-                    );
-                Navigator.pop(context);
-              }
-            },
-            child: const Text('Update'),
-          ),
-        ],
-      ),
-    );
-  }
-
-  void _showDeleteDialog(BuildContext context) {
-    showDialog(
-      context: context,
-      builder: (context) => AlertDialog(
-        title: const Text('Delete Agent'),
-        content: Text('Are you sure you want to delete "${agent.name}"?'),
-        actions: [
-          TextButton(
-            onPressed: () => Navigator.pop(context),
-            child: const Text('Cancel'),
-          ),
-          ElevatedButton(
-            onPressed: () {
-              context.read<AgentBloc>().add(DeleteAgent(agent.id));
-              Navigator.pop(context);
-            },
-            style: ElevatedButton.styleFrom(
-              backgroundColor: Colors.red,
-              foregroundColor: Colors.white,
-            ),
-            child: const Text('Delete'),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class _TemplateCard extends StatelessWidget {
-  final AgentModel template;
-
-  const _TemplateCard({required this.template});
-
-  @override
-  Widget build(BuildContext context) {
-    return Card(
-      margin: const EdgeInsets.only(bottom: 12),
-      child: ListTile(
-        leading: Container(
-          padding: const EdgeInsets.all(8),
-          decoration: BoxDecoration(
-            color: Theme.of(context).colorScheme.secondaryContainer,
-            borderRadius: BorderRadius.circular(8),
-          ),
-          child: Icon(
-            Icons.auto_awesome,
-            color: Theme.of(context).colorScheme.onSecondaryContainer,
-          ),
-        ),
-        title: Text(
-          template.name,
-          style: Theme.of(context).textTheme.titleMedium,
-        ),
-        subtitle: Text(
-          template.description,
-          style: Theme.of(context).textTheme.bodyMedium,
-        ),
-        trailing: ElevatedButton(
-          onPressed: () {
-            context.read<AgentBloc>().add(ImportAgent(template));
-            ScaffoldMessenger.of(context).showSnackBar(
-              SnackBar(
-                content: Text('Created agent from "${template.name}" template'),
-              ),
-            );
-          },
-          child: const Text('Use Template'),
-        ),
-        onTap: () => _showTemplateDetails(context),
-      ),
-    );
-  }
-
-  void _showTemplateDetails(BuildContext context) {
-    showDialog(
-      context: context,
-      builder: (context) => AlertDialog(
-        title: Text('${template.name} Template'),
-        content: SingleChildScrollView(
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            mainAxisSize: MainAxisSize.min,
-            children: [
-              Text(
-                'Description',
-                style: Theme.of(context).textTheme.titleSmall,
-              ),
-              const SizedBox(height: 4),
-              Text(template.description),
-              const SizedBox(height: 16),
-              Text(
-                'System Prompt',
-                style: Theme.of(context).textTheme.titleSmall,
-              ),
-              const SizedBox(height: 4),
-              Text(template.systemPrompt),
-            ],
-          ),
-        ),
-        actions: [
-          TextButton(
-            onPressed: () => Navigator.pop(context),
-            child: const Text('Close'),
-          ),
-          ElevatedButton(
-            onPressed: () {
-              context.read<AgentBloc>().add(ImportAgent(template));
-              Navigator.pop(context);
-              ScaffoldMessenger.of(context).showSnackBar(
-                SnackBar(
-                  content:
-                      Text('Created agent from "${template.name}" template'),
-                ),
-              );
-            },
-            child: const Text('Use Template'),
-          ),
-        ],
-      ),
+      builder: (context) => const CreateAgentDialog(),
     );
   }
 }
